create a patch queue from patchqueue branch
added patches:
add-field-revdate-to-asciidoc-files-for-manpages.patch
libcoap-1.pc.in-adding-missed-API-version.patch

Without the first patch the package libcoap-1-0-bin isn't reproducible
because the timestamp within the manpages are set by the system time
while creating. Using the timestamp from the changelog file is
preventing this.
The second patch fixes the missing version on the linkerflags for
pkgconfig.